Factor some logic out of the cli module, which was getting unwieldy.
